+++ src/ui/Scrollbar.h
@@ -21,6 +21,6 @@ namespace ui
 	export_ TWO_UI_EXPORT ScrollbarStyles& scrollbar_styles();
 
 	export_ TWO_UI_EXPORT bool overflow(Frame& frame, Frame& content, Axis dim);
-	export_ TWO_UI_EXPORT Widget& scrollbar(Widget& parent, Frame& frame, Frame& content, Axis dim, v2<size_t> grid_index = { 0, 0 });
+	export_ TWO_UI_EXPORT Widget& scrollbar(Widget& parent, Frame& frame, Frame& content, Axis dim, v2<uint> grid_index = { 0, 0 });
 }
 }
